+++ app/src/androidTest/java/woowacourse/movie/reservation/ReservationDetailActivityTest.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.activity
+package woowacourse.movie.reservation
 
 import androidx.test.core.app.ActivityScenario
 import androidx.test.espresso.Espresso.onView
@@ -15,6 +15,8 @@ import org.junit.Test
 import org.junit.runner.RunWith
 import woowacourse.movie.R
 import woowacourse.movie.TestFixture.moviesFirstItem
+import woowacourse.movie.home.ReservationHomeActivity
+import woowacourse.movie.reservation.detail.ReservationDetailActivity
 
 @RunWith(AndroidJUnit4::class)
 class ReservationDetailActivityTest {
