+++ app/src/androidTest/java/woowacourse/movie/reservation/ReservationFinishedActivityTest.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.activity
+package woowacourse.movie.reservation
 
 import android.content.Intent
 import androidx.test.core.app.ActivityScenario
@@ -14,6 +14,8 @@ import org.junit.Rule
 import org.junit.Test
 import woowacourse.movie.R
 import woowacourse.movie.model.Ticket
+import woowacourse.movie.reservation.detail.ReservationDetailActivity
+import woowacourse.movie.reservation.finished.ReservationFinishedActivity
 
 class ReservationFinishedActivityTest {
     @get:Rule
