+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.Movie"
         tools:targetApi="31">
         <activity
-            android:name=".activity.ReservationHomeActivity"
+            android:name=".home.ReservationHomeActivity"
             android:exported="true"
             android:screenOrientation="portrait">
             <intent-filter>
@@ -23,13 +23,13 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".activity.ReservationFinishedActivity"
-            android:parentActivityName=".activity.ReservationHomeActivity"
+            android:name=".reservation.finished.ReservationFinishedActivity"
+            android:parentActivityName=".home.ReservationHomeActivity"
             android:exported="false"
             android:screenOrientation="portrait" />
         <activity
-            android:name=".activity.ReservationDetailActivity"
-            android:parentActivityName=".activity.ReservationHomeActivity"
+            android:name=".reservation.detail.ReservationDetailActivity"
+            android:parentActivityName=".home.ReservationHomeActivity"
             android:exported="false"
             android:screenOrientation="portrait" />
     </application>
