+++ app/src/main/java/woowacourse/movie/home/ReservationHomeActivity.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.activity
+package woowacourse.movie.home
 
 import android.content.Intent
 import android.os.Bundle
@@ -6,8 +6,7 @@ import android.widget.ListView
 import androidx.appcompat.app.AppCompatActivity
 import woowacourse.movie.R
 import woowacourse.movie.adapter.MovieCatalogAdapter
-import woowacourse.movie.presenter.ReservationHomePresenter
-import woowacourse.movie.view.ReservationHomeContract
+import woowacourse.movie.reservation.detail.ReservationDetailActivity
 
 class ReservationHomeActivity : AppCompatActivity(), ReservationHomeContract {
     private val movies: ListView by lazy { findViewById(R.id.list_view_reservation_home) }
@@ -17,9 +16,10 @@ class ReservationHomeActivity : AppCompatActivity(), ReservationHomeContract {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_reservation_home)
 
-        movies.adapter = MovieCatalogAdapter(this, reservationHomePresenter.movies) { movie ->
-            reservationHomePresenter.deliverMovie(movie.id)
-        }
+        movies.adapter =
+            MovieCatalogAdapter(this, reservationHomePresenter.movies) { movie ->
+                reservationHomePresenter.deliverMovie(movie.id)
+            }
     }
 
     override fun moveToReservationDetail(movieId: Int) {
