+++ app/src/main/java/woowacourse/movie/reservation/detail/ReservationDetailContract.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view
+package woowacourse.movie.reservation.detail
 
 import woowacourse.movie.model.Movie
 import woowacourse.movie.model.Ticket
