+++ app/src/main/java/woowacourse/movie/reservation/detail/ReservationDetailPresenter.kt
@@ -1,11 +1,10 @@
-package woowacourse.movie.presenter
+package woowacourse.movie.reservation.detail
 
 import woowacourse.movie.model.ChangeTicketCountResult
 import woowacourse.movie.model.Failure
 import woowacourse.movie.model.Movies
 import woowacourse.movie.model.Success
 import woowacourse.movie.model.Ticket
-import woowacourse.movie.view.ReservationDetailContract
 
 class ReservationDetailPresenter(
     private val contract: ReservationDetailContract,
