+++ app/src/main/java/woowacourse/movie/reservation/finished/ReservationFinishedActivity.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.activity
+package woowacourse.movie.reservation.finished
 
 import android.content.Context
 import android.content.Intent
@@ -8,11 +8,10 @@ import android.widget.TextView
 import androidx.activity.OnBackPressedCallback
 import androidx.appcompat.app.AppCompatActivity
 import woowacourse.movie.R
-import woowacourse.movie.activity.ReservationDetailActivity.Companion.DEFAULT_MOVIE_ID
+import woowacourse.movie.home.ReservationHomeActivity
 import woowacourse.movie.model.Movie
 import woowacourse.movie.model.Ticket
-import woowacourse.movie.presenter.ReservationFinishedPresenter
-import woowacourse.movie.view.ReservationFinishedContract
+import woowacourse.movie.reservation.detail.ReservationDetailActivity.Companion.DEFAULT_MOVIE_ID
 import java.io.Serializable
 import java.text.DecimalFormat
